requirements.txt
openai
requests
python-dotenv
hdbcli
